Permit customization of application environment
Specifically, make it easy to customize the environment of a specific
application at build time, without having to specify all settings in a
single file. While rebar app_vars_file may work in a similar fashion,
having a minimal file that overrides only the environment settings is
ideal for the purpose of customizing the application build. Besides,
there seems to be a lack of a good use of app_vars_file.

make ct passes