Ignore the example file in incoming directory
As an easy reminder on how to name the directories within the incoming
directory, allow a placeholder file that is explicitly ignored by the
incoming script. Otherwise the script will warn about the unknown file,
which is good behavior, but would be annoying considering the idea is
for the script to be run frequently.