Show thumbnails on the tagged page
Show the thumbnails for the assets with matching tags in a table 3 columns
wide. Include the original file name and date of the asset.

Moved the broken image back to the backend, where it is needed for returning
a valid binary even if generating a thumbnail fails (it is cached so we
do not try regenerating a thumbnail over and over again).