Add an admin page for uncommon actions
A new admin page that allows the user to easily rename a tag across all
matching documents. Same with the location field. In a similar fashion, find
all documents with a certain tag, remove the tag and set the tag as the
location (only if the document does not have a location value already).

Add a "by_location" CouchDB view for the rename-location function to use.
This will later be used to display assets with a certain location value,
just as with tags.

mix test passes